+++ backend/app/routes.py
@@ -1,0 +1,7 @@
+
+from app import app
+
+
+@app.route('/message')
+def get_current_time():
+    return {'message': 'Hello from backend!'}
