+++ backend/assets/images/bigger_pixel_test.png
